+++ src/interrogate/visit.py
@@ -179,6 +179,17 @@ class CoverageVisitor(ast.NodeVisitor):
                     return True
         return False
 
+    def _has_setters(self, node):
+        """Detect if node has property get/setter decorators."""
+        if not hasattr(node, "decorator_list"):
+            return False
+
+        for dec in node.decorator_list:
+            if hasattr(dec, "attr"):
+                if dec.attr == "setter":
+                    return True
+        return False
+
     def _is_func_ignored(self, node):
         """Should the AST visitor ignore this func/method node."""
         is_init = node.name == "__init__"
@@ -190,12 +201,15 @@ class CoverageVisitor(ast.NodeVisitor):
             ]
         )
         has_property_decorators = self._has_property_decorators(node)
+        has_setters = self._has_setters(node)
 
         if self.config.ignore_init_method and is_init:
             return True
         if self.config.ignore_magic and is_magic:
             return True
         if self.config.ignore_property_decorators and has_property_decorators:
+            return True
+        if self.config.ignore_property_setters and has_setters:
             return True
 
         return self._is_ignored_common(node)